refactor: move IDConflictError type and its constructor out of mapstorage.go to a separate file
